revamp controller and supervisor logic
